requirements.txt
gspread
selenium
webdriver-manager
python-dotenv
